linkedin: refactor
